LLD-3-advanced-software-engineering/message-queues-email-service/email_notification_project/requirements.txt
django>=4.2
djangorestframework>=3.14
kafka-python>=2.0.2
python-dotenv>=1.0.0
